+++ src/Sitecore.FakeDb/Data/Engines/DataStorage.cs
@@ -44,8 +44,6 @@ namespace Sitecore.FakeDb.Data.Engines
 
     private readonly IDictionary<ID, DbItem> fakeItems;
 
-    private readonly IDictionary<ID, DbTemplate> fakeTemplates;
-
     private readonly IDictionary<Guid, Stream> blobs;
 
     public DataStorage(Database database)
@@ -60,7 +58,6 @@ namespace Sitecore.FakeDb.Data.Engines
     internal DataStorage()
     {
       this.fakeItems = new Dictionary<ID, DbItem>();
-      this.fakeTemplates = new Dictionary<ID, DbTemplate>();
       this.blobs = new Dictionary<Guid, Stream>();
     }
 
@@ -74,11 +71,6 @@ namespace Sitecore.FakeDb.Data.Engines
       get { return this.fakeItems; }
     }
 
-    public IDictionary<ID, DbTemplate> FakeTemplates
-    {
-      get { return this.fakeTemplates; }
-    }
-
     public IDictionary<Guid, Stream> Blobs
     {
       get { return this.blobs; }
@@ -88,33 +80,43 @@ namespace Sitecore.FakeDb.Data.Engines
     {
       Assert.ArgumentNotNull(item, "item");
 
+      var loading = item is IDsDbItem;
       if (item as DbTemplate != null)
       {
         var template = (DbTemplate)item;
+
+        if (!loading)
+        {
+          this.AssertNoTemplateExists(template);
+        }
 
         if (template is IDsDbItem)
         {
           CorePipeline.Run("loadDsDbTemplate", new DsItemLoadingArgs(template as IDsDbItem, this));
         }
-        else
-        {
-          Assert.ArgumentCondition(!this.FakeTemplates.ContainsKey(template.ID), "template", "A template with the same id has already been added.");
-        }
 
-        // TODO: Should be Add()?
-        this.FakeTemplates[template.ID] = template;
-        this.Database.Engines.TemplateEngine.Reset();
       }
 
-      if (item is IDsDbItem)
+      // TODO: Combine the two pipelines below.
+      if (loading)
       {
         CorePipeline.Run("loadDsDbItem", new DsItemLoadingArgs(item as IDsDbItem, this));
       }
 
       CorePipeline.Run("addDbItem", new AddDbItemArgs(item, this));
 
-      // TODO: Should be Add()?
+      if (!loading)
+      {
+        this.AssertNoItemExists(item);
+      }
+
       this.FakeItems[item.ID] = item;
+
+      if (item as DbTemplate != null)
+      {
+        this.Database.Engines.TemplateEngine.Reset();
+      }
+
       foreach (var child in item.Children)
       {
         child.ParentID = item.ID;
@@ -132,9 +134,12 @@ namespace Sitecore.FakeDb.Data.Engines
 
     public virtual DbTemplate GetFakeTemplate(ID templateId)
     {
-      Assert.ArgumentCondition(!ID.IsNullOrEmpty(templateId), "templateId", "Value cannot be null.");
+      return this.FakeItems.ContainsKey(templateId) ? this.FakeItems[templateId] as DbTemplate : null;
+    }
 
-      return this.FakeTemplates.ContainsKey(templateId) ? this.FakeTemplates[templateId] : null;
+    public IEnumerable<DbTemplate> GetFakeTemplates()
+    {
+      return this.FakeItems.Values.OfType<DbTemplate>();
     }
 
     public virtual Item GetSitecoreItem(ID itemId)
@@ -251,7 +256,7 @@ namespace Sitecore.FakeDb.Data.Engines
       {
         foreach (var id in fakeTemplate.BaseIDs)
         {
-          this.AddFieldsFromTemplate(allFields, fakeItem, this.fakeTemplates[id], language, version);
+          this.AddFieldsFromTemplate(allFields, fakeItem, this.GetFakeTemplate(id), language, version);
         }
       }
     }
@@ -272,33 +277,37 @@ namespace Sitecore.FakeDb.Data.Engines
 
     protected void FillDefaultFakeTemplates()
     {
-      this.FakeTemplates.Add(TemplateIdSitecore, new DbTemplate("Sitecore", new TemplateID(TemplateIdSitecore)) { new DbField(FieldIDs.Security) });
-      this.FakeTemplates.Add(TemplateIDs.MainSection, new DbTemplate("Main Section", TemplateIDs.MainSection));
+      this.FakeItems.Add(TemplateIdSitecore, new DbTemplate("Sitecore", new TemplateID(TemplateIdSitecore)) { new DbField(FieldIDs.Security) });
+      this.FakeItems.Add(TemplateIDs.MainSection, new DbTemplate("Main Section", TemplateIDs.MainSection));
 
-      this.FakeTemplates.Add(TemplateIDs.Template, new DbTemplate(TemplateItemName, TemplateIDs.Template) { new DbField(FieldIDs.BaseTemplate) });
-      this.FakeTemplates.Add(TemplateIDs.Folder, new DbTemplate(FolderItemName, TemplateIDs.Folder));
+      this.FakeItems.Add(
+        TemplateIDs.Template,
+        new DbTemplate(TemplateItemName, TemplateIDs.Template)
+          {
+            ParentID = ItemIDs.TemplateRoot,
+            FullPath = "/sitecore/templates/template",
+            Fields = { new DbField(FieldIDs.BaseTemplate) }
+          });
 
-      var standardTemplate = new DbTemplate(TemplateIDs.StandardTemplate)
-                               {
-                                 new DbField(FieldIDs.BaseTemplate) { Shared = true },
+      this.FakeItems.Add(TemplateIDs.Folder, new DbTemplate(FolderItemName, TemplateIDs.Folder));
 
-                                 new DbField(FieldIDs.Lock) { Shared = true },
-                                 new DbField(FieldIDs.Security) { Shared = true },
-                                 
-                                 new DbField(FieldIDs.Created),
-                                 new DbField(FieldIDs.CreatedBy),
-                                 new DbField(FieldIDs.Updated),
-                                 new DbField(FieldIDs.UpdatedBy),
-                                 new DbField(FieldIDs.Revision),
-
-                                 new DbField(FieldIDs.LayoutField),
-
-                                 new DbField(FieldIDs.DisplayName),
-                                 new DbField(FieldIDs.Hidden),
-                                 new DbField(FieldIDs.ReadOnly)
-                               };
-
-      this.FakeTemplates.Add(standardTemplate.ID, standardTemplate);
+      this.FakeItems.Add(
+        TemplateIDs.StandardTemplate,
+        new DbTemplate(TemplateIDs.StandardTemplate)
+          {
+            new DbField(FieldIDs.BaseTemplate) { Shared = true },
+            new DbField(FieldIDs.Lock) { Shared = true },
+            new DbField(FieldIDs.Security) { Shared = true },
+            new DbField(FieldIDs.Created),
+            new DbField(FieldIDs.CreatedBy),
+            new DbField(FieldIDs.Updated),
+            new DbField(FieldIDs.UpdatedBy),
+            new DbField(FieldIDs.Revision),
+            new DbField(FieldIDs.LayoutField),
+            new DbField(FieldIDs.DisplayName),
+            new DbField(FieldIDs.Hidden),
+            new DbField(FieldIDs.ReadOnly)
+          });
     }
 
     protected void FillDefaultFakeItems()
@@ -318,10 +327,30 @@ namespace Sitecore.FakeDb.Data.Engines
       this.FakeItems[ItemIDs.RootID].Add(this.FakeItems[ItemIDs.MediaLibraryRoot]);
 
       // TODO: Move 'Template' item to proper directory to correspond Sitecore structure.
-      this.FakeItems.Add(TemplateIDs.Template, new DbItem(TemplateItemName, TemplateIDs.Template, TemplateIDs.Template) { ParentID = ItemIDs.TemplateRoot, FullPath = "/sitecore/templates/template" });
       this.FakeItems.Add(TemplateIDs.TemplateSection, new DbItem(TemplateSectionItemName, TemplateIDs.TemplateSection, TemplateIDs.Template) { ParentID = ItemIDs.TemplateRoot, FullPath = "/sitecore/templates/template section" });
       this.FakeItems.Add(TemplateIDs.TemplateField, new DbItem(TemplateFieldItemName, TemplateIDs.TemplateField, TemplateIDs.Template) { ParentID = ItemIDs.TemplateRoot, FullPath = "/sitecore/templates/template field" });
       this.FakeItems.Add(TemplateIDs.BranchTemplate, new DbItem(BranchItemName, TemplateIDs.BranchTemplate, TemplateIDs.Template) { ParentID = ItemIDs.TemplateRoot, FullPath = "/sitecore/templates/branch" });
+    }
+
+    private void AssertNoTemplateExists(DbItem item)
+    {
+      this.AssertDoesNotExists(item, "A template with the same id has already been added ('{0}', '{1}').");
+    }
+
+    private void AssertNoItemExists(DbItem item)
+    {
+      this.AssertDoesNotExists(item, "An item with the same id has already been added ('{0}', '{1}').");
+    }
+
+    private void AssertDoesNotExists(DbItem item, string format)
+    {
+      if (!this.FakeItems.ContainsKey(item.ID))
+      {
+        return;
+      }
+
+      var message = string.Format(format, item.ID, item.FullPath ?? item.Name);
+      throw new InvalidOperationException(message);
     }
   }
 }